add tests modified
