--- tests/code.py
@@ -1,6 +1,0 @@
-def up_first(msg):
-    """Делает первую букву строки заглавной."""
-    if msg:
-        return msg[0].upper() + msg[1:]
-    else:
-        return msg
