+++ tests/test_widget.py
@@ -101,3 +101,5 @@ class TestGetDateFunction(unittest.TestCase):
 
 if __name__ == '__main__':
     unittest.main()
+
+
